[Refactor] Adapter Package 추가, Model Package 이동, Adapter Refactor
