+++ app/src/main/java/com/example/hangman/data/model/Info.kt
@@ -1,3 +1,3 @@
-package com.example.hangman
+package com.example.hangman.data.model
 
 data class Info(var roomTitle: String, var count: String)
